fix tests after merge conflict resolution
